app/src/main/java/com/example/feature_mainpage_gui/ui/screens/counter/CounterScreenViewModel.kt
package com.example.feature_mainpage_gui.ui.screens.counter

import androidx.compose.runtime.getValue
import androidx.compose.runtime.mutableIntStateOf
import androidx.compose.runtime.mutableStateOf
import androidx.compose.runtime.setValue
import androidx.lifecycle.ViewModel
import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.StateFlow

class CounterScreenViewModel : ViewModel() {

    private val _count = MutableStateFlow(0)
    val count: StateFlow<Int> = _count

    fun increment() {
        _count.value += 1
    }

    fun decrement() {
        _count.value = if (_count.value == 0) 0 else _count.value - 1
    }

    fun reset() {
        _count.value = 0
    }
}